Ensure that all traces get reported to the agent
Previously, traces were added to the current trace set at the beginning.
And after reporting, the trace set was cleared. So if the reporting
interval is 1 second, any traces taking more than 1 second (or, really,
any trace that crosses that boundary) would be incomplete since it would
be reported and removed from the trace set before completion.