Extract span creation and finalization to methods
The Crystal compiler inlines blocks into methods that yield, which
causes `Tracer#trace` to be compiled dozens/hundreds of times, leading
to super long compile times when the method is that large. Extracting
the creation and finalization of the span should speed up compilation
quite a bit. This also allows us to reactivate Armature template
render instrumentation.